requirements.txt
beautifulsoup4==4.4.1
bs4==0.0.1
requests==2.9.1
Django==1.9.6
lxml==3.6.0
django-constance[database]
pytz==2016.4